streaming refactor
