+++ www/app/api/chat/honcho/route.ts
@@ -1,0 +1,24 @@
+import { getUserData } from '@/utils/ai';
+import { honcho } from '@/utils/honcho';
+import { NextRequest, NextResponse } from 'next/server';
+
+export async function POST(req: NextRequest) {
+  const { message, conversationId } = await req.json();
+
+  const userData = await getUserData();
+
+  if (!userData) {
+    return new Response('Unauthorized', { status: 401 });
+  }
+
+  const { appId, userId } = userData;
+
+  const dialecticQuery = await honcho.apps.users.sessions.chat(
+    appId,
+    userId,
+    conversationId,
+    { queries: message }
+  );
+
+  return NextResponse.json({ content: dialecticQuery.content });
+}
